+++ server/api/index.js
@@ -2,10 +2,10 @@ const router = require('express').Router();
 module.exports = router;
 
 router.use('/s3', require('./s3-visionApi.js'));
-router.use('/spotify', require('./spotify'));
+router.use('/spotify', require('./spotify-tracks'));
 
 router.use((req, res, next) => {
-	const error = new Error('Not Found');
-	error.status = 404;
-	next(error);
+  const error = new Error('Not Found');
+  error.status = 404;
+  next(error);
 });
